vehicle_posting_site_backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.0

# CORS
django-cors-headers>=4.0.0

# Authentication
PyJWT>=2.8.0

# Image handling (for vehicle images)
Pillow>=10.0.0

# AI & Verification
openai>=1.12.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
